Cad Cliente Ok, Busca quase lá

+++ PetShop/Model/Pessoa.cs
@@ -10,16 +10,17 @@ namespace PetShop.Model
     {
         private int codCli;
         private string nome;
-        private int cpf;
+        private long cpf;
         private string cep;
         private string endereco;
         private string cidade;
         private string numero;
         private string telefone;
 
+
         public int CodCli { get => codCli; set => codCli = value; }
         public string Nome { get => nome; set => nome = value; }
-        public int Cpf { get => cpf; set => cpf = value; }
+        public long Cpf { get => cpf; set => cpf = value; }
         public string Cep { get => cep; set => cep = value; }
         public string Endereco { get => endereco; set => endereco = value; }
         public string Cidade { get => cidade; set => cidade = value; }
